unify all reduce ops
