+++ ttnn/cpp/ttnn/operations/experimental/ccl/all_reduce_async/all_reduce_async_pybind.cpp
@@ -27,9 +27,9 @@ void bind_all_reduce_async(pybind11::module& module, const ccl_operation_t& oper
         ttnn::pybind_overload_t{
             [](const ccl_operation_t& self,
                const ttnn::Tensor& input_tensor,
-               const GlobalSemaphore& from_remote_multi_device_global_semaphore,
-               const GlobalSemaphore& to_remote_multi_device_global_semaphore,
-               const GlobalSemaphore& gather_multi_device_global_semaphore,
+               const uint32_t num_devices,
+               const std::vector<GlobalSemaphore>& rs_global_semaphores,
+               const std::vector<GlobalSemaphore>& ag_global_semaphores,
                ttnn::operations::reduction::ReduceType math_op,
                const ttnn::MemoryConfig& memory_config,
                ttnn::ccl::Topology topology,
@@ -37,9 +37,9 @@ void bind_all_reduce_async(pybind11::module& module, const ccl_operation_t& oper
                std::optional<tt::tt_metal::SubDeviceId> worker_subdevice_id_opt) -> ttnn::Tensor {
                 return self(
                     input_tensor,
-                    from_remote_multi_device_global_semaphore,
-                    to_remote_multi_device_global_semaphore,
-                    gather_multi_device_global_semaphore,
+                    num_devices,
+                    rs_global_semaphores,
+                    ag_global_semaphores,
                     math_op,
                     memory_config,
                     topology,
@@ -47,9 +47,9 @@ void bind_all_reduce_async(pybind11::module& module, const ccl_operation_t& oper
                     worker_subdevice_id_opt);
             },
             py::arg("input_tensor"),
-            py::arg("from_remote_multi_device_global_semaphore"),
-            py::arg("to_remote_multi_device_global_semaphore"),
-            py::arg("gather_multi_device_global_semaphore"),
+            py::arg("num_devices"),
+            py::arg("rs_global_semaphores"),
+            py::arg("ag_global_semaphores"),
             py::arg("math_op"),
             py::kw_only(),
             py::arg("memory_config") = std::nullopt,
@@ -62,9 +62,8 @@ void bind_all_reduce_async(pybind11::module& module, const ccl_operation_t& oper
                const ttnn::Tensor& input_tensor,
                const uint32_t cluster_axis,
                const MeshDevice& mesh_device,
-               const GlobalSemaphore& from_remote_multi_device_global_semaphore,
-               const GlobalSemaphore& to_remote_multi_device_global_semaphore,
-               const GlobalSemaphore& gather_multi_device_global_semaphore,
+               const std::vector<GlobalSemaphore>& rs_global_semaphores,
+               const std::vector<GlobalSemaphore>& ag_global_semaphores,
                ttnn::operations::reduction::ReduceType math_op,
                const ttnn::MemoryConfig& memory_config,
                ttnn::ccl::Topology topology,
@@ -74,9 +73,8 @@ void bind_all_reduce_async(pybind11::module& module, const ccl_operation_t& oper
                     input_tensor,
                     cluster_axis,
                     mesh_device,
-                    from_remote_multi_device_global_semaphore,
-                    to_remote_multi_device_global_semaphore,
-                    gather_multi_device_global_semaphore,
+                    rs_global_semaphores,
+                    ag_global_semaphores,
                     math_op,
                     memory_config,
                     topology,
@@ -86,9 +84,8 @@ void bind_all_reduce_async(pybind11::module& module, const ccl_operation_t& oper
             py::arg("input_tensor"),
             py::arg("cluster_axis"),
             py::arg("mesh_device"),
-            py::arg("from_remote_multi_device_global_semaphore"),
-            py::arg("to_remote_multi_device_global_semaphore"),
-            py::arg("gather_multi_device_global_semaphore"),
+            py::arg("rs_global_semaphores"),
+            py::arg("ag_global_semaphores"),
             py::arg("math_op"),
             py::kw_only(),
             py::arg("memory_config") = std::nullopt,
