--- ttnn/cpp/ttnn/operations/experimental/ccl/all_reduce_minimal_async/all_reduce_minimal_async_pybind.hpp
@@ -1,13 +1,0 @@
-// SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
-//
-// SPDX-License-Identifier: Apache-2.0
-
-#pragma once
-
-#include "ttnn-pybind/pybind_fwd.hpp"
-
-namespace ttnn::operations::experimental::ccl {
-
-void py_bind_all_reduce_minimal_async(pybind11::module& module);
-
-}  // namespace ttnn::operations::experimental::ccl
